+++ CHANGELOG.md
@@ -18,10 +18,13 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Fixed
 
+- Fix branch exists git function ([82])
 - Creation of new repositories made more robust ([79])
 
 
+[82]: https://github.com/openlawlibrary/taf/pull/82
 [79]: https://github.com/openlawlibrary/taf/pull/79
+
 
 ## [0.1.8] - 11/12/2019